Let PRE elements be replaced with internal markdown
PRE-elements will preserve line spaces and formatting of the
code inside for the user, so it is counter intuitive to mix
PRE tags with markdown.

However, when using inline markdown in this world of precompilerr,
linting and automated formatting the PRE-element is a safe-zone
for formatting rules.

As PRE would normally not be used with inline markdown formatting
I suggest impelemnting this breaking change (major version has
been bumped).